+++ json-decode.go
@@ -126,12 +126,15 @@ func (cdc *Codec) _decodeReflectJSON(bz []byte, info *TypeInfo, rv reflect.Value
 	case reflect.Struct:
 		err = cdc.decodeReflectJSONStruct(bz, info, rv, opts)
 
+	case reflect.Map:
+		err = cdc.decodeReflectJSONMap(bz, info, rv, opts)
+
 	//----------------------------------------
 	// Signed, Unsigned
 
 	case reflect.Int64, reflect.Int32, reflect.Int16, reflect.Int8, reflect.Int,
 		reflect.Uint64, reflect.Uint32, reflect.Uint16, reflect.Uint8, reflect.Uint:
-		err = invokeStdlibJSONUnmarshal(bz, info, rv, opts)
+		err = invokeStdlibJSONUnmarshal(bz, rv, opts)
 
 	//----------------------------------------
 	// Misc
@@ -142,7 +145,7 @@ func (cdc *Codec) _decodeReflectJSON(bz []byte, info *TypeInfo, rv reflect.Value
 		}
 		fallthrough
 	case reflect.Bool, reflect.String:
-		err = invokeStdlibJSONUnmarshal(bz, info, rv, opts)
+		err = invokeStdlibJSONUnmarshal(bz, rv, opts)
 
 	//----------------------------------------
 	// Default
@@ -154,7 +157,7 @@ func (cdc *Codec) _decodeReflectJSON(bz []byte, info *TypeInfo, rv reflect.Value
 	return
 }
 
-func invokeStdlibJSONUnmarshal(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) error {
+func invokeStdlibJSONUnmarshal(bz []byte, rv reflect.Value, opts FieldOptions) error {
 	if !rv.CanAddr() && rv.Kind() != reflect.Ptr {
 		panic("rv not addressable nor pointer")
 	}
@@ -419,6 +422,60 @@ func (cdc *Codec) decodeReflectJSONStruct(bz []byte, info *TypeInfo, rv reflect.
 	return nil
 }
 
+// CONTRACT: rv.CanAddr() is true.
+func (cdc *Codec) decodeReflectJSONMap(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) (err error) {
+	if !rv.CanAddr() {
+		panic("rv not addressable")
+	}
+	if printLog {
+		fmt.Println("(d) decodeReflectJSONMap")
+		defer func() {
+			fmt.Printf("(d) -> err: %v\n", err)
+		}()
+	}
+
+	// Map all the fields(keys) to their blobs/bytes.
+	// NOTE: In decodeReflectBinaryMap, we don't need to do this,
+	// since fields are encoded in order.
+	var rawMap = make(map[string]json.RawMessage)
+	err = json.Unmarshal(bz, &rawMap)
+	if err != nil {
+		return
+	}
+
+	var krt = rv.Type().Key()
+	if krt.Kind() != reflect.String {
+		err = fmt.Errorf("decodeReflectJSONMap: key type must be string") // TODO also support []byte and maybe others
+		return
+	}
+	var vinfo *TypeInfo
+	vinfo, err = cdc.getTypeInfo_wlock(rv.Type().Elem())
+	if err != nil {
+		return
+	}
+
+	var mrv = reflect.MakeMapWithSize(rv.Type(), len(rawMap))
+	for key, valueBytes := range rawMap {
+
+		// Get map value rv.
+		vrv := reflect.New(mrv.Type().Elem()).Elem()
+
+		// Decode valueBytes into vrv.
+		err = cdc.decodeReflectJSON(valueBytes, vinfo, vrv, opts)
+		if err != nil {
+			return
+		}
+
+		// And set.
+		krv := reflect.New(reflect.TypeOf("")).Elem()
+		krv.SetString(key)
+		mrv.SetMapIndex(krv, vrv)
+	}
+	rv.Set(mrv)
+
+	return nil
+}
+
 //----------------------------------------
 // Misc.
 
@@ -460,6 +517,10 @@ func decodeDisfixJSON(bz []byte) (df DisfixBytes, data []byte, err error) {
 	}
 
 	// Get data.
+	if len(dfw.Data) == 0 {
+		err = errors.New("Disfix JSON wrapper should have non-empty value field")
+		return
+	}
 	data = dfw.Data
 	return
 }